Automation Testing/requirements.txt
pytest==8.2.4
requests==2.32.3
pytest-html==4.1.1
pyyaml==6.0.1
